Bumped WP to 6.3.1, updated all deprecations and warnings in code, fixed gutenberg bugs

+++ wp-content/plugins/better-wp-security/core/lib/class-itsec-lib-user-activity.php
@@ -30,7 +30,7 @@ final class ITSEC_Lib_User_Activity {
 			return false;
 		}
 
-		return get_user_meta( $user_id, 'itsec_user_activity_last_seen', true );
+		return (int) get_user_meta( $user_id, 'itsec_user_activity_last_seen', true );
 	}
 
 	public function identify_user() {
